organized contexts and added logic for setting avatars
