+++ frontend/src/index.js
@@ -3,9 +3,9 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import {UserInfoProvider} from './UserContext';
-import {PartiesInfoProvider} from './PartiesContext';
-import {HomeRenderProvider} from './HomeRenderContext';
+import {UserInfoProvider} from './context/UserContext';
+import {PartiesInfoProvider} from './context/PartiesContext';
+import {HomeRenderProvider} from './context/HomeRenderContext';
 
 ReactDOM.render(
   <React.StrictMode>
